refactor: move slideshow properties
