+++ frontend/src/stores/presentation.js
@@ -1,5 +1,6 @@
 import { createResource } from 'frappe-ui'
 import { ref } from 'vue'
+import { changeSlide } from './slideActions'
 
 const presentationId = ref('')
 const presentation = createResource({
@@ -13,4 +14,28 @@ const presentationList = createResource({
 	auto: true,
 })
 
-export { presentationId, presentation, presentationList }
+const inSlideShow = ref(false)
+const applyReverseTransition = ref(false)
+
+const startSlideShow = async () => {
+	if (!presentation.data) await presentation.reload()
+	await changeSlide(0)
+	const elem = document.querySelector('.slideContainer')
+
+	if (elem.requestFullscreen) {
+		elem.requestFullscreen()
+	} else if (elem.webkitRequestFullscreen) {
+		elem.webkitRequestFullscreen()
+	} else if (elem.msRequestFullscreen) {
+		elem.msRequestFullscreen()
+	}
+}
+
+export {
+	presentationId,
+	presentation,
+	presentationList,
+	inSlideShow,
+	applyReverseTransition,
+	startSlideShow,
+}
